[REF] mail: introduce suggestion service
It even has its own folder!
That's what you get when promoted to a discuss feature :)

(it cleans the code too, but that's secondary)
