+++ addons/mail/static/src/new/composer/composer.js
@@ -22,7 +22,7 @@ import { FileUploader } from "@web/views/fields/file_handler";
 import { Typing } from "./typing";
 import { useDebounced } from "@web/core/utils/timing";
 import { browser } from "@web/core/browser/browser";
-import { useSuggestion } from "./composer_hook";
+import { useSuggestion } from "../suggestion/suggestion_hook";
 
 import { _t } from "@web/core/l10n/translation";
