+++ addons/mail/static/src/new/core/messaging.js
@@ -1270,69 +1270,6 @@ export class Messaging {
         return partners;
     }
 
-    searchChannelCommand(cleanedSearchTerm, thread, sort) {
-        if (!["chat", "channel", "group"].includes(thread.type)) {
-            // channel commands are channel specific
-            return [[]];
-        }
-        const commands = commandRegistry
-            .getEntries()
-            .filter(([name, command]) => {
-                if (!cleanTerm(name).includes(cleanedSearchTerm)) {
-                    return false;
-                }
-                if (command.channel_types) {
-                    return command.channel_types.includes(thread.type);
-                }
-                return true;
-            })
-            .map(([name, command]) => {
-                return {
-                    channel_types: command.channel_types,
-                    help: command.help,
-                    id: command.id,
-                    name,
-                };
-            });
-        const sortFunc = (a, b) => {
-            const isATypeSpecific = a.channel_types;
-            const isBTypeSpecific = b.channel_types;
-            if (isATypeSpecific && !isBTypeSpecific) {
-                return -1;
-            }
-            if (!isATypeSpecific && isBTypeSpecific) {
-                return 1;
-            }
-            const cleanedAName = cleanTerm(a.name || "");
-            const cleanedBName = cleanTerm(b.name || "");
-            if (
-                cleanedAName.startsWith(cleanedSearchTerm) &&
-                !cleanedBName.startsWith(cleanedSearchTerm)
-            ) {
-                return -1;
-            }
-            if (
-                !cleanedAName.startsWith(cleanedSearchTerm) &&
-                cleanedBName.startsWith(cleanedSearchTerm)
-            ) {
-                return 1;
-            }
-            if (cleanedAName < cleanedBName) {
-                return -1;
-            }
-            if (cleanedAName > cleanedBName) {
-                return 1;
-            }
-            return a.id - b.id;
-        };
-        return [
-            {
-                type: "ChannelCommand",
-                suggestions: sort ? commands.sort(sortFunc) : commands,
-            },
-        ];
-    }
-
     async leaveChannel(channel) {
         await this.orm.call("mail.channel", "action_unfollow", [channel.id]);
         channel.remove();
@@ -1415,95 +1352,6 @@ export class Messaging {
             thread.customName = name;
             await this.orm.call("mail.channel", "channel_set_custom_name", [[thread.id]], { name });
         }
-    }
-
-    /**
-     * Returns suggestions that match the given search term from specified type.
-     *
-     * @param {Object} [param0={}]
-     * @param {String} [param0.delimiter] can be one one of the following: ["@", ":", "#", "/"]
-     * @param {String} [param0.term]
-     * @param {Object} [options={}]
-     * @param {Integer} [options.thread] prioritize and/or restrict
-     *  result in the context of given thread
-     * @returns {[mainSuggestion[], extraSuggestion[]]}
-     */
-    searchSuggestions({ delimiter, term }, { thread } = {}, sort = false) {
-        const cleanedSearchTerm = cleanTerm(term);
-        switch (delimiter) {
-            case "@": {
-                return Partner.searchSuggestions(this.state, cleanedSearchTerm, thread, sort);
-            }
-            case ":":
-                return CannedResponse.searchSuggestions(this.state, cleanedSearchTerm, sort);
-            case "#":
-                return Thread.searchSuggestions(this.state, cleanedSearchTerm, thread, sort);
-            case "/":
-                return this.searchChannelCommand(cleanedSearchTerm, thread, sort);
-        }
-        return [
-            {
-                type: undefined,
-                suggestions: [],
-            },
-            {
-                type: undefined,
-                suggestions: [],
-            },
-        ];
-    }
-
-    async fetchSuggestions({ delimiter, term }, { thread } = {}) {
-        const cleanedSearchTerm = cleanTerm(term);
-        switch (delimiter) {
-            case "@": {
-                this.fetchPartners(cleanedSearchTerm, thread);
-                break;
-            }
-            case ":":
-                break;
-            case "#":
-                this.fetchThreads(cleanedSearchTerm);
-                break;
-            case "/":
-                break;
-        }
-    }
-
-    async fetchPartners(term, thread) {
-        const kwargs = { search: term };
-        const isNonPublicChannel =
-            thread &&
-            (thread.type === "group" ||
-                thread.type === "chat" ||
-                (thread.type === "channel" && thread.serverData.group_based_subscription));
-        if (isNonPublicChannel) {
-            kwargs.channel_id = thread.id;
-        }
-        const suggestedPartners = await this.orm.call(
-            "res.partner",
-            "get_mention_suggestions",
-            [],
-            kwargs
-        );
-        suggestedPartners.map((data) => {
-            Partner.insert(this.state, data);
-        });
-    }
-
-    async fetchThreads(term) {
-        const suggestedThreads = await this.orm.call(
-            "mail.channel",
-            "get_mention_suggestions",
-            [],
-            { search: term }
-        );
-        suggestedThreads.map((data) => {
-            Thread.insert(this.state, {
-                model: "mail.channel",
-                ...data,
-            });
-        });
     }
 
     async notifyThreadDescriptionToServer(thread, description) {
